Initial Release
Grit is a traditional card game I developed.
The rules are available here:
http://www.stfj.net/index2.php?project=art/2013/Grit.pdf

This repository is a bluetooth based iOS version of the game.
It’s very rough and doesn’t explain the rules or how to play at all,
but it does work really well. If you like it you can download it and
play it.

The images for the cards are from a google image search and I
definitely DO NOT have the rights to use them.

Aside from the card art, the code and game game itself is licensed
under the Attribution-ShareAlike 4.0 license -
http://creativecommons.org/licenses/by-sa/4.0/

It should compile with OpenFrameworks 0.7.4 iOS (
http://www.openframeworks.cc/versions/v0.7.4/of_v0.7.4_ios_release.zip
).

It’s provided on an AS IS basis and I can’t offer any support for it.

The attribution required for the license is: “Based on a game design by
Zach Gage - stfj.net”